+++ inst/examples/colourInput/app.R
@@ -2,7 +2,7 @@ library(shiny)
 library(colourpicker)
 
 share <- list(
-  title = "colourInput",
+  title = "colourInput()",
   url = "http://daattali.com/shiny/colourInput/",
   image = "http://daattali.com/shiny/img/colourinput.png",
   description = "An input control that allows users to select colours in Shiny apps",