+++ NEWS.md
@@ -1,3 +1,11 @@
+# colourpicker 0.1.1
+
+2016-08-15
+
+- upgrade to newer version of JS library that fixed bugs with new jquery
+- add `runExample()` function to run the example shiny app
+
+
 # colourpicker 0.1
 
 2016-08-11